+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased] - ReleaseDate
 
+### Breaking
+
+- Remove `preview_templates` config field. Previews will always be rendered now
+
 ### Fixed
 
 - Fix TUI crash when using an empty select list